# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kodementor Api Documentation",
        "description": "Kodementor Api Documentation",
        "contact": {
            "name": "Vijay Rana",
            "email": "info@kodementor.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/addCard": {
            "post": {
                "tags": [
                    "Card"
                ],
                "summary": "add card",
                "operationId": "addCard",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "example": [
                                    {
                                        "card_owner": "vannsann",
                                        "card_number": "123456789",
                                        "cvv": "321",
                                        "expire_date": "2/2025"
                                    },
                                    {
                                        "card_owner": "alex john",
                                        "card_number": "987654321",
                                        "cvv": "123",
                                        "expire_date": "2/2025"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User registered successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/getCard": {
            "get": {
                "tags": [
                    "Card"
                ],
                "summary": "get card by user",
                "operationId": "getCard",
                "responses": {
                    "201": {
                        "description": "User registered successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/getPopularFilter": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "get category ",
                "operationId": "getPopularFilter",
                "responses": {
                    "201": {
                        "description": "Category successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/addOrder": {
            "post": {
                "tags": [
                    "Order"
                ],
                "summary": "add Order",
                "operationId": "addOrder",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "example": {
                                    "items": [
                                        1,
                                        2,
                                        3,
                                        4
                                    ],
                                    "grant_qty": 10,
                                    "grant_total": 20,
                                    "address": {
                                        "name": "up",
                                        "country": "Cambodia",
                                        "city": "Phnom penh",
                                        "phone number": "097878765",
                                        "address": "1st 225 psa depo2"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User registered successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/getProduct": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "get products",
                "operationId": "getProduct",
                "parameters": [
                    {
                        "name": "brand_id",
                        "in": "query",
                        "description": "brand_id",
                        "required": false,
                        "schema": {
                            "type": "int"
                        },
                        "example": 1
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "search Index",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "shoe"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page Index",
                        "required": false,
                        "schema": {
                            "type": "int"
                        },
                        "example": 1
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "perPage Index",
                        "required": false,
                        "schema": {
                            "type": "int"
                        },
                        "example": 10
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User registered successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/getVisitPlaceByProvince": {
            "get": {
                "tags": [
                    "Visit Place"
                ],
                "summary": "get visit place by province",
                "operationId": "getVisitPlaceByprovince",
                "parameters": [
                    {
                        "name": "province",
                        "in": "query",
                        "description": "province",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Kampot"
                    },
                    {
                        "name": "cagetories",
                        "in": "query",
                        "description": "cagetories",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "farm"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page Index",
                        "required": false,
                        "schema": {
                            "type": "int"
                        },
                        "example": 1
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "perPage Index",
                        "required": false,
                        "schema": {
                            "type": "int"
                        },
                        "example": 10
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User registered successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "plainTextToken",
                "scheme": "bearer"
            }
        }
    }
}